+++ packages/didactic-settings/src/didactic/settings/__init__.py
@@ -27,7 +27,7 @@ from didactic.settings._settings import (
     Settings,
 )
 
-__version__ = "0.7.4"
+__version__ = "0.7.5"
 
 __all__ = [
     "CliSource",