+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.5] - 2026-06-16
+
 ### Added
 
 - The inbound synthesiser (``didactic.synthesis``: ``model_from_spec``,